[Akshay] | Adds src and moves file into different folders
